medico envia sinal de vida ao balcao

+++ src/balcao.c
@@ -67,7 +67,7 @@ void handle_sig(int signo, siginfo_t *info, void *context){
                 medicos[i].especialidade[0] = '\0';
             }
         }
-        alarm(10);
+        alarm(6);
     }
 }
 
@@ -134,7 +134,7 @@ int main(int argc, char **argv, char **envp) {
     sigaction(SIGINT, &action, NULL); // avisar que vamos sair
     sigaction(SIGUSR2, &action, NULL); // Avisaram que sairam
     sigaction(SIGALRM, &action, NULL); // Avisaram que sairam
-    alarm(10);
+    alarm(6);
     // definição de variáveis
     int numMedicos = 0;
     Utente utenteNovo;
@@ -202,8 +202,8 @@ int main(int argc, char **argv, char **envp) {
         if (nfd == 0) {
             fflush(stdout);
             continue;
-        }
-        /*if (nfd == -1) {
+        }/*
+        if (nfd == -1) {
             perror("\nerro no select");
             close(s_c_fifo_fd);
             unlink(SERVER_FIFO_CLIENTES);
@@ -225,7 +225,7 @@ int main(int argc, char **argv, char **envp) {
                 int i;
                 printf("Especialistas\n\n");
                 for (i=0;i<maxMedicos;i++){
-                    if(medicos[i].especialidade[0] != '\0'){
+                    if(medicos[i].pid != 0){
                         printf("Nome: %s\tEspecialidade: %s\n", medicos[i].nome, medicos[i].especialidade);
                     }
                 }
@@ -240,7 +240,9 @@ int main(int argc, char **argv, char **envp) {
                     }
                 }
                 for (i=0; i<maxMedicos; i++) {
-                    kill(medicos[i].pid, SIGUSR2);
+                    if(medicos[i].pid != 0){
+                        kill(medicos[i].pid, SIGUSR2);
+                    }
                 }
 
                 return 0;
@@ -261,7 +263,7 @@ int main(int argc, char **argv, char **envp) {
                     for (i = 0;i< maxMedicos; i++) {
                         if(strcmp(resto, medicos[i].nome) == 0) {
                             kill(medicos[i].pid, SIGUSR2);
-                            medicos[i].especialidade[0] = '\0';
+                            medicos[i].pid = 0;
                             printf("Especialista eliminado!\n");
                         }
                     }
@@ -370,11 +372,13 @@ int main(int argc, char **argv, char **envp) {
                     strcpy(medicos[numMedicos].nome, buffer.nome);
                     strcpy(medicos[numMedicos].especialidade, buffer.msg);
                     medicos[numMedicos].pid = buffer.pid;
+                    medicos[numMedicos].ts = buffer.ts;
                     numMedicos++;
                     msg_med.pid = getpid();
                     msg_med.tipo = 1;
                     write(fifo_medico, &msg_med, sizeof(B_M));
                     close(fifo_medico);
+                    printf("medico adicionado!\n");
                 }
             }
             // medico terminou consulta
@@ -394,7 +398,7 @@ int main(int argc, char **argv, char **envp) {
                 for(i=0; i<maxMedicos; i++) {
                     if(medicos[i].pid == buffer.pid) {
                         medicos[i].ts = buffer.ts;
-                        //printf("Timestamp do medico %d: %d\n", buffer.pid, buffer.ts);
+                        printf("Timestamp do medico %d: %d\n", buffer.pid, buffer.ts);
                     }
                 }
             }
